libcxx: updated to 8.0.0:
What’s New in Libc++ 8.0.0?

API Changes

Building libc++ for Mac OSX 10.6 is not supported anymore.
Starting with LLVM 8.0.0, users that wish to link together translation units built with different versions of libc++’s headers into the same final linked image MUST define the _LIBCPP_HIDE_FROM_ABI_PER_TU macro to 1 when building those translation units. Not defining _LIBCPP_HIDE_FROM_ABI_PER_TU to 1 and linking translation units built with different versions of libc++’s headers together may lead to ODR violations and ABI issues. On the flipside, code size improvements should be expected for everyone not defining the macro.
Starting with LLVM 8.0.0, std::dynarray has been removed from the library. std::dynarray was a feature proposed for C++14 that was pulled from the Standard at the last minute and was never standardized. Since there are no plans to standardize this facility it is being removed.
Starting with LLVM 8.0.0, std::bad_array_length has been removed from the library. std::bad_array_length was a feature proposed for C++14 alongside std::dynarray, but it never actually made it into the C++ Standard. There are no plans to standardize this feature at this time. Formally speaking, this removal constitutes an ABI break because the symbols were shipped in the shared library. However, on macOS systems, the feature was not usable because it was hidden behind availability annotations. We do not expect any actual breakage to happen from this change.