add "--without-ns" to x11 configs for OSX
on OSX the configure script will select the "nextstep" window system
unless you explicitly tell it not to:

    % bmake configure |& egrep 'PKG_OPTIONS.emacs|What window system should Emacs use'
            PKG_OPTIONS.emacs = x11
      What window system should Emacs use?                    nextstep
    %

leaving pkgsrc in a bad state where it thinks it built an x11 version,
but instead emacs built a nextstep version.

add "--without-ns" to the x11 config option for OSX.  this
should not impact other platforms.  (this diff corresponds to
the change i did in options.mk 1.4 for editors/emacs25)