Update to 0.29.1
Changelog:
What's new in notmuch 0.29.1
=========================

Build
-----

Fix for installation failure with `configure --without-emacs`.

What is notmuch
===============
Notmuch is a system for indexing, searching, reading, and tagging
large collections of email messages in maildir or mh format. It uses
the Xapian library to provide fast, full-text search with a convenient
search syntax.